+++ src/pages/index.pug
@@ -1,0 +1,9 @@
+ul
+    li
+        a(href='login/login.pug') Login
+    li
+        a(href='signin/signin.pug') Signin
+    li
+        a(href='chats/chats.pug') Chats
+    li
+        a(href='profile/profile.pug') Profile
